se renderiza a partir de posiciones relativas

+++ common/include/creatorMessage.h
@@ -5,22 +5,21 @@
 
 #include "../include/entity.h"
 
-/* TODO: es Message (doble s). */
 class CreatorMessage{
 private:
 	ENTITY id_class;
 	uint32_t id_object;
-	float posx;
-	float posy;
+	int32_t posx;
+	int32_t posy;
 	uint32_t dir;
 public:
-	explicit CreatorMessage(ENTITY id_class,	uint32_t id_object,
-	float posx,	float posy,uint32_t dir);
+	explicit CreatorMessage(ENTITY id_class, uint32_t id_object,
+	int32_t posx, int32_t posy, uint32_t dir);
 	~CreatorMessage();
 	ENTITY getIdClass() const;
 	uint32_t getIdObject() const;
-	float getPosX() const;
-	float getPosY() const;
+	int32_t getPosX() const;
+	int32_t getPosY() const;
 	uint32_t getDir() const;
 };
